Modul 6.4 4th

+++ kodilla-testing/src/main/java/com/kodilla/testing/shape/Shape.java
@@ -4,6 +4,6 @@ public interface Shape {
 
     String getShapeName();
 
-    int getField();
+    double getField();
 
 }
